Moved NPevents function to NPCalendar and changed imports
